correct sinatra skeleton nightmare
