+++ Gemfile
@@ -1,5 +1,5 @@
-source 'https://rubygems.org'
 ruby '2.0.0'
+source 'https://rubygems.org'
 
 # PostgreSQL driver
 gem 'pg'
@@ -7,25 +7,24 @@ gem 'pg'
 # Sinatra driver
 gem 'sinatra'
 gem 'sinatra-contrib'
-gem 'bcrypt'
+
 gem 'sinatra-flash'
 gem 'sinatra-redirect-with-flash'
-gem 'pry'
 
-gem 'activesupport', '~>4.2.0'
-gem 'activerecord', '~>4.2.0'
+# Use Thin for our web server
+gem 'thin'
+
+gem 'activesupport', '~>4.1'
+gem 'activerecord', '~>4.1'
 
 gem 'rake'
+
+gem 'pry'
+gem 'bcrypt'
 gem 'faker'
+
 gem 'shotgun'
 
 group :test do
-  gem 'shoulda-matchers'
-  gem 'rack-test'
-  gem 'rspec', '~>3.0'
-  gem 'capybara'
-end
-
-group :test, :development do
-  gem 'factory_girl'
+  gem 'rspec'
 end
