+++ Rakefile
@@ -1,4 +1,6 @@
 require 'rake'
+require 'rspec/core/rake_task'
+
 
 require ::File.expand_path('../config/environment', __FILE__)
 
@@ -84,19 +86,16 @@ namespace :generate do
 end
 
 namespace :db do
-  desc "Drop, create, and migrate the database"
-  task :reset => [:drop, :create, :migrate]
-
-  desc "Create the databases at #{DB_NAME}"
+  desc "Create the database at #{DB_NAME}"
   task :create do
-    puts "Creating development and test databases if they don't exist..."
-    system("createdb #{APP_NAME}_development && createdb #{APP_NAME}_test")
+    puts "Creating database #{DB_NAME} if it doesn't exist..."
+    exec("createdb #{DB_NAME}")
   end
 
   desc "Drop the database at #{DB_NAME}"
   task :drop do
-    puts "Dropping development and test databases..."
-    system("dropdb #{APP_NAME}_development && dropdb #{APP_NAME}_test")
+    puts "Dropping database #{DB_NAME}..."
+    exec("dropdb #{DB_NAME}")
   end
 
   desc "Migrate the database (options: VERSION=x, VERBOSE=false, SCOPE=blog)."
@@ -108,28 +107,21 @@ namespace :db do
     end
   end
 
-  desc "rollback your migration--use STEP=number to step back multiple times"
-  task :rollback do
-    step = (ENV['STEP'] || 1).to_i
-    ActiveRecord::Migrator.rollback('db/migrate', step)
-    Rake::Task['db:version'].invoke if Rake::Task['db:version']
-  end
-
   desc "Populate the database with dummy data by running db/seeds.rb"
   task :seed do
     require APP_ROOT.join('db', 'seeds.rb')
   end
 
+  desc "rollback your migration--use STEPS=number to step back multiple times"
+  task :rollback do
+    steps = (ENV['STEPS'] || 1).to_i
+    ActiveRecord::Migrator.rollback('db/migrate', steps)
+    Rake::Task['db:version'].invoke if Rake::Task['db:version']
+  end
+
   desc "Returns the current schema version number"
   task :version do
     puts "Current version: #{ActiveRecord::Migrator.current_version}"
-  end
-
-  namespace :test do
-    desc "Migrate test database"
-    task :prepare do
-      system "rake db:migrate RACK_ENV=test"
-    end
   end
 end
 
@@ -138,14 +130,7 @@ task "console" do
   exec "pry -r./config/environment"
 end
 
+desc "Run the specs"
+RSpec::Core::RakeTask.new(:spec)
 
-# In a production environment like Heroku, RSpec might not
-# be available.  To handle this, rescue the LoadError.
-# https://devcenter.heroku.com/articles/getting-started-with-ruby-o#runtime-dependencies-on-development-test-gems
-begin
-  require 'rspec/core/rake_task'
-  RSpec::Core::RakeTask.new(:spec)
-rescue LoadError
-end
-
-task :default  => :spec
+task :default  => :specs
