--- app/controllers/sessions.rb
@@ -1,19 +1,0 @@
-get '/login' do
-  erb :'users/login'
-end
-
-post '/login' do
-  user = User.find_by(username: params[:username]).try(:authenticate, params[:password])
-  if user
-    session[:user_id] = user.id
-    redirect '/'
-  else
-    #add flash error
-    redirect '/fail'
-  end
-end
-
-delete '/logout' do
-  session.clear
-  redirect '/'
-end
