--- app/controllers/welcome.rb
@@ -1,4 +1,0 @@
-get '/' do
-  @user =
-  erb :'/home'
-end
