--- app/helpers/current_user.rb
@@ -1,3 +1,0 @@
-def current_user
-  user = User.find_by(id: session[:user_id])
-end
