--- app/helpers/logged_in.rb
@@ -1,3 +1,0 @@
-def logged_in?
-  return true if session[:user_id]
-end
