--- app/models/choice.rb
@@ -1,6 +1,0 @@
-class Choice < ActiveRecord::Base
-  belongs_to :question
-  has_many :responses
-
-  validates_presence_of :body
-end
