--- app/models/question.rb
@@ -1,6 +1,0 @@
-class Question < ActiveRecord::Base
-  belongs_to :survey
-  has_many :choices
-
-  validates_presence_of :body
-end
