--- app/models/response.rb
@@ -1,5 +1,0 @@
-class Response < ActiveRecord::Base
-  belongs_to :choice
-  belongs_to :user
-  belongs_to :taken_survey
-end
