--- app/models/taken_survey.rb
@@ -1,6 +1,0 @@
-class TakenSurvey < ActiveRecord::Base
-  belongs_to :survey
-  belongs_to :taker, class_name: :User
-
-  has_many :responses, through: :taker
-end
