--- app/views/_header.erb
@@ -1,13 +1,0 @@
-<header>
-  <a href="/">Home</a>
-  <%if logged_in?%>
-    <a href='/users/<%=session[:user_id]%>'>Profile</a>
-    <form action='/logout' method='post'/>
-      <input type="hidden" name="_method" value="DELETE">
-      <input type='submit' value='Log Out'/>
-    </form>
-  <%else%>
-    <a href="/login">Login</a>
-    <a href="/signup">Sign Up</a>
-  <%end%>
-</header>
