--- app/views/questions/show.erb
@@ -1,13 +1,0 @@
-<h1 class="title"><%=@survey.title%></h1>
-
-<h3 class="question"><%=@question.body%></h3>
-
-<form action="/responses" method="post">
-  <input type="hidden" name="taken_survey_id" value="<%=@taken_survey.id%>">
-  <ul>
-  <% @question.choices.each do |choice| %>
-    <li><input type="radio" name="choice_id" value="<%=choice.body%>"><%=choice.body%></li>
-  <% end %>
-  </ul>
-  <input type="submit" value="Submit response">
-</form>
