--- app/views/surveys/show.erb
@@ -1,1 +1,0 @@
-<h1>Statistics for <%=@survey.title%></h1>
