--- app/views/taken_surveys/new.erb
@@ -1,7 +1,0 @@
-<h1 class="title"><%=@survey.title%></h1>
-
-<form action="/takensurveys" method="post" >
-  <input type="hidden" name="id" value="<%= @survey.id %>" />
-  <input type="submit" value="Start survey!" />
-</form>
-
