--- app/views/users/_form.erb
@@ -1,6 +1,0 @@
-  <label for='username'>Username: </label>
-  <input id='username' type='text' name='user[username]'/>
-  <label for='password'>Password: </label>
-  <input id='password' type='password' name='user[password]'/>
-  <label for='username'>Email: </label>
-  <input id='username' type='text' name='user[email]'/>
