--- app/views/users/edit.erb
@@ -1,7 +1,0 @@
-<div id='profile-edit-form'>
-  <form action='/users' method='POST'>
-    <input type="hidden" name="_method" value="PUT" />
-    <%= erb :'users/_form' %>
-    <input type='submit' value='Update Profile'/>
-  </form>
-</div>
