--- app/views/users/login.erb
@@ -1,9 +1,0 @@
-<div id='Login-form'>
-  <form action='/login' method='POST'>
-    <label for='username'>Enter Username: </label>
-    <input id='username' type='text' name='username'/>
-    <label for='password'>Enter Password: </label>
-    <input id='password' type='password' name='password'/>
-    <input type='submit' value='Login!'/>
-  </form>
-</div>
