--- app/views/users/logout.erb
@@ -1,6 +1,0 @@
-<p>
-  <form action='/logout' method='POST'>
-    <input type="hidden" name="_method" value="DELETE" />
-    <input type='submit' value='Log Out'/>
-  </form>
-</p>
