--- app/views/users/new.erb
@@ -1,6 +1,0 @@
-<div id='Signup-form'>
-  <form action='/users' method='POST'>
-    <%= erb :'users/_form' %>
-    <input type='submit' value='Sign Up'/>
-  </form>
-</div>
