+++ config/environment.rb
@@ -10,17 +10,20 @@ require 'rubygems'
 
 require 'uri'
 require 'pathname'
-require 'bcrypt'
-require 'sinatra/flash'
-require 'sinatra/redirect_with_flash'
-require 'pry'
 
 require 'pg'
 require 'active_record'
 require 'logger'
 
+require 'pry'
+require 'bcrypt'
+require 'faker'
+
 require 'sinatra'
 require "sinatra/reloader" if development?
+
+require 'sinatra/flash'
+require 'sinatra/redirect_with_flash'
 
 require 'erb'
 
@@ -28,18 +31,6 @@ require 'erb'
 APP_ROOT = Pathname.new(File.expand_path('../../', __FILE__))
 
 APP_NAME = APP_ROOT.basename.to_s
-
-configure do
-  # By default, Sinatra assumes that the root is the file that calls the configure block.
-  # Since this is not the case for us, we set it manually.
-  set :root, APP_ROOT.to_path
-  # See: http://www.sinatrarb.com/faq.html#sessions
-  enable :sessions
-  set :session_secret, ENV['SESSION_SECRET'] || 'this is a secret shhhhh'
-
-  # Set the views to
-  set :views, File.join(Sinatra::Application.root, "app", "views")
-end
 
 # Set up the controllers and helpers
 Dir[APP_ROOT.join('app', 'controllers', '*.rb')].each { |file| require file }
