--- db/migrate/20150828143300_create_questions.rb
@@ -1,9 +1,0 @@
-class CreateQuestions < ActiveRecord::Migration
-  def change
-  	create_table :questions do |t|
-  		t.string :body, null:false, limit: 256
-  		t.references :survey
-  		t.timestamps
-  	end	
-  end
-end
