--- db/migrate/20150828143313_create_taken_surveys.rb
@@ -1,9 +1,0 @@
-class CreateTakenSurveys < ActiveRecord::Migration
-  def change
-  	create_table :taken_surveys do |t|
-  		t.references :survey 
-  		t.references :taker
-  		t.timestamps
-  	end 	
-  end
-end
