--- db/migrate/20150828143324_create_responses.rb
@@ -1,9 +1,0 @@
-class CreateResponses < ActiveRecord::Migration
-  def change
-  	create_table :responses do |t|
-  		t.references :choice
-  		t.references :user
-  		t.timestamps
-  	end	
-  end
-end
