+++ db/migrate/20150828210500_create_users.rb
@@ -1,0 +1,10 @@
+class CreateUsers < ActiveRecord::Migration
+  def change
+    create_table :users do |t|
+      t.string :username, null:false
+      t.string :email, null:false
+      t.string :password_digest
+      t.timestamps
+    end
+  end
+end
