--- public/js/application.js
@@ -1,1 +1,0 @@
-
